clean: Rename utils to analysis, AnalogLattice to AnalogTypeLattice
