+++ src/oqd_core/frontend/analysis/utils.py
