+++ tests/test_frontend/test_frontend_analog.py
@@ -16,9 +16,9 @@ import pytest
 
 from oqd_core.frontend.analog.AnalogCircuitAST import parse_analog
 from oqd_core.frontend.analog.cfg import AnalogCFGBuilder
-from oqd_core.frontend.utils.utils import SCCAnalysis
 from oqd_core.frontend.analog.serialize import serialize_analog
 from oqd_core.frontend.analog.type_checker import AnalogTypeChecker, AnalogTypeError
+from oqd_core.frontend.analysis.utils import SCCAnalysis
 from oqd_core.interface.analog import (
     Access,
     AnalogCircuit,
